chore: make services run more deterministically
- The original approach to run the `start` script is to run it under `nohup` and place it in background, redirecting
its stdout and stderr somewhere, as well as the nginx, php-fpm and redis-server services. But this way we may
unfortunately lost all the useful information while starting up the services, like a config error causing it fail to
start, or a permission problem.
- In this commit, I removed all the `nohup` commands, `&` operators, and also the output redirection. Were there any
problem during service startup, we might easily find it. It helps us to reduce meaningless debugging time and
concentrate on developement.
- The reason why we can do this is that all these services are daemonisable, they can be run in background via
corresponding configs.
- This commit also slightly improves php-fpm performance by substituting http listening with socket listening, b/c the
nginx and php-fpm services are run in the same container, they can communicate directly avoiding networking costs.

Signed-off-by: Eugene Cheng <hsxfjames@gmail.com>